Move e2e page objects to page-object-models/ to match client-core
Move feature page-object classes from playwright/utils/planning and
playwright/utils/contacts to playwright/page-object-models, mirroring
superdesk-client-core's layout. Shared helper functions and generic base
UI wrappers stay in utils/. Update all spec imports accordingly.
