+++ e2e/playwright/assignments/reassign_assignment.spec.ts
@@ -1,7 +1,7 @@
 import {test, expect} from '@playwright/test';
 
 import {setup, login, waitForPageLoad, SubNavBar, changeWorkspace, Modal} from '../utils/common';
-import {PlanningList, PlanningEditor, AssignmentEditor, AssignmentPreview} from '../utils/planning';
+import {PlanningList, PlanningEditor, AssignmentEditor, AssignmentPreview} from '../page-object-models/planning';
 import {getMenuItem} from '../utils/common';
 
 test.describe('Planning.Assignment: reassign assignment', () => {
