+++ e2e/playwright/page-object-models/planning/planningPreview.ts
@@ -1,6 +1,6 @@
 import {Page, Locator} from '@playwright/test';
 
-import {ActionMenu} from '../common';
+import {ActionMenu} from '../../utils/common';
 
 export class PlanningPreview {
     page: Page;
